+++ src/main/java/gift/common/client/kakao/KakaoProperties.java
@@ -1,4 +1,4 @@
-package gift.client.kakao;
+package gift.common.client.kakao;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
